[tb/update-locations] Move expression util into widget folder
