+++ packages/perseus-core/src/index.ts
@@ -39,5 +39,4 @@ export type * from "./widgets/logic-export.types";
 
 export {default as getOrdererPublicWidgetOptions} from "./widgets/orderer/orderer-util";
 export {default as getCategorizerPublicWidgetOptions} from "./widgets/categorizer/categorizer-util";
-
-export {default as getExpressionPublicWidgetOptions} from "./utils/expression-util";
+export {default as getExpressionPublicWidgetOptions} from "./widgets/expression/expression-util";
